conference_selection/main/requirements.txt
PyMuPDF>=1.22.0
pytesseract>=0.3.10
Pillow>=9.0.1
requests>=2.25.0
pathway>=0.8.0
transformers>=4.30.0
google-generativeai>=0.2.0
